backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.3.2
